Initial implementation of seam carving algorithm
There are a few tests but the code needs to be reorganizd as it is
pretty messy right now and needs to be more modular. Better test cases
as well need to be made as the code is reoganized.

I also have plans to do some horizonal lowest energy removal as well, as
the image looks weird only removing horizontally.